feat: implement changing current tree and install shadcn textarea.

+++ binary_search_tree/src/components/TreeForm.tsx
@@ -12,6 +12,7 @@ import {
   FormMessage,
 } from '@/components/ui/form';
 import { Input } from './ui/input';
+import { Textarea } from './ui/textarea';
 import { Tree } from '@/logic/Tree';
 
 type TreeFormProps = {
@@ -23,7 +24,7 @@ type TreeFormProps = {
 };
 
 const formSchema = z.object({
-  nodes: z.array(z.number()),
+  nodes: z.string(),
   node: z.string(),
   deleteNode: z.string(),
 });
@@ -32,14 +33,39 @@ export const TreeForm = (props: TreeFormProps) => {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      nodes: [],
+      nodes: '',
       node: '',
       deleteNode: '',
     },
   });
 
+  function onUpdateNodes(values: z.infer<typeof formSchema>) {
+    if (!values.nodes) {
+      form.setError('nodes', {
+        type: 'min',
+        message: 'Required. String of numbers only, separated by spaces.',
+      });
+    }
+    const treeArray = values.nodes.split(' ').map((str) => Number(str));
+    if (treeArray.some((num) => isNaN(num))) {
+      form.setError('nodes', {
+        type: 'custom',
+        message: 'String of numbers only, separated by spaces.',
+      });
+      return;
+    }
+    props.updateTree(values.nodes.split(' ').map((str) => Number(str)));
+  }
+
   function onInsertNode(values: z.infer<typeof formSchema>) {
     const currentKey = parseInt(values.node);
+    if (isNaN(currentKey)) {
+      form.setError('node', {
+        type: 'min',
+        message: 'Required. Number strings only.',
+      });
+      return;
+    }
     if (props.tree.find(currentKey)?.key === currentKey) {
       form.setError('node', {
         type: 'custom',
@@ -48,17 +74,61 @@ export const TreeForm = (props: TreeFormProps) => {
       return;
     }
     props.insertNode(parseInt(values.node));
-    form.reset();
+    form.reset({ node: '' });
   }
 
   function onDeleteNode(values: z.infer<typeof formSchema>) {
+    const currentKey = parseInt(values.deleteNode);
+    if (isNaN(currentKey)) {
+      form.setError('deleteNode', {
+        type: 'min',
+        message: 'Required. Number strings only.',
+      });
+      return;
+    }
     props.deleteNode(parseInt(values.deleteNode));
-    form.reset();
+    form.reset({ deleteNode: '' });
   }
 
   return (
     <Form {...form}>
       <form className="space-y-8 px-2">
+        <Button
+          type="button"
+          className="w-full sm:w-max"
+          onClick={props.rebalance}
+        >
+          Rebalance Tree
+        </Button>
+        <div>
+          <FormField
+            control={form.control}
+            name="nodes"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Update tree nodes</FormLabel>
+                <FormControl>
+                  <Textarea
+                    placeholder="String of numbers separated by spaces."
+                    className="resize-none"
+                    {...field}
+                  />
+                </FormControl>
+                <FormDescription>
+                  A collection of nodes that will be in the tree.
+                </FormDescription>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <Button
+            type="button"
+            className="w-full sm:w-max mt-2"
+            onClick={form.handleSubmit(onUpdateNodes)}
+          >
+            Update Tree
+          </Button>
+        </div>
         <div>
           <FormField
             control={form.control}
@@ -78,7 +148,7 @@ export const TreeForm = (props: TreeFormProps) => {
           />
           <Button
             type="button"
-            className="w-full sm:w-max"
+            className="w-full sm:w-max mt-2"
             onClick={form.handleSubmit(onInsertNode)}
           >
             Insert
@@ -103,16 +173,12 @@ export const TreeForm = (props: TreeFormProps) => {
           />
           <Button
             type="button"
-            className="w-full sm:w-max"
+            className="w-full sm:w-max mt-2"
             onClick={form.handleSubmit(onDeleteNode)}
           >
             Delete
           </Button>
         </div>
-
-        <Button type="button" onClick={props.rebalance}>
-          Rebalance
-        </Button>
       </form>
     </Form>
   );
